+++ builder.go
@@ -6,7 +6,6 @@ import (
 	"io/ioutil"
 	"mime"
 	"net/mail"
-	"net/smtp"
 	"net/textproto"
 	"os"
 	"path/filepath"
@@ -307,9 +306,8 @@ func (p MailBuilder) Build() (*Part, error) {
 	return root, nil
 }
 
-// SendWithReturnAddress encodes the message and sends it via the SMTP server specified by addr
-// and from. Send uses net/smtp.SendMail, and accepts the same authentication parameters.
-func (p MailBuilder) SendWithReturnAddress(addr, from string, a smtp.Auth) error {
+// SendWithReversePath encodes the message and sends it via the specified Sender.
+func (p MailBuilder) SendWithReversePath(sender Sender, from string) error {
 	buf := &bytes.Buffer{}
 	root, err := p.Build()
 	if err != nil {
@@ -329,13 +327,13 @@ func (p MailBuilder) SendWithReturnAddress(addr, from string, a smtp.Auth) error
 	for _, a := range p.bcc {
 		recips = append(recips, a.Address)
 	}
-	return smtp.SendMail(addr, a, from, recips, buf.Bytes())
+	return sender.Send(from, recips, buf.Bytes())
 }
 
-// Send encodes the message and sends it via the SMTP server specified by addr. Send uses
-// net/smtp.SendMail, and accepts the same authentication parameters.
-func (p MailBuilder) Send(addr string, a smtp.Auth) error {
-	return p.SendWithReturnAddress(addr, p.from.Address, a)
+// Send encodes the message and sends it via the specified Sender, using the address provided to
+// `From()` as the reverse-path.
+func (p MailBuilder) Send(sender Sender) error {
+	return p.SendWithReversePath(sender, p.from.Address)
 }
 
 // Equals uses the reflect package to test two MailBuilder structs for equality, primarily for unit